pyproject.toml
[project]
name = "mcpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
